Rename api_server2 module to apiserver
